+++ ion/gnn/__init__.py
@@ -1,15 +1,17 @@
-from jax.ops import segment_max, segment_min, segment_prod, segment_sum
-
 from .gat import GATConv, GATv2Conv
-from .gcn import GCNConv
+from .gcn import GCNConv, GraphConv
 from .gin import GINConv
 from .ops import (
     add_self_loops,
     batch_graphs,
     max_pool,
     mean_pool,
+    segment_max,
     segment_mean,
+    segment_min,
+    segment_prod,
     segment_softmax,
+    segment_sum,
     sum_pool,
 )
 from .sage import SAGEConv
@@ -18,6 +20,7 @@ __all__ = [
     "GATConv",
     "GATv2Conv",
     "GCNConv",
+    "GraphConv",
     "GINConv",
     "SAGEConv",
     "add_self_loops",